Department membership delete bug
When an employee is removed from a department membership, but still belongs to schedules in that department, it will break the front end. Added some extra logic to test for this, but did not change backend, which probably should be done:

Probably should remove employee from all schedules of that department (But not delete schedules themselves.)

Actually, this should especially be done because it will still calculate those hours and costs while showing no employee assigned.